refactor: extract services mixin from api/state.py
